Factor out loadFromFilename().
